refactoring of attestation components
