+++ src/constants/eas.ts
@@ -1,0 +1,4 @@
+// src/constants/eas.ts
+export const EAS_CONTRACT_ADDRESS = "0x4200000000000000000000000000000000000021";
+export const SCHEMA_UID = "0xb763e62d940bed6f527dd82418e146a904e62a297b8fa765c9b3e1f0bc6fdd68";
+export const SCHEMA_DEFINITION = 'string chain_id,string tags_json';
